// packages/admin-sdk/src/zod/generated/emails/EmailStockLocation.ts
// This file is auto-generated. Do not edit directly.
import { z } from 'zod';

export const EmailStockLocationSchema = z.object({
  id: z.string(),
  name: z.string(),
  address1: z.string().nullable(),
  city: z.string().nullable(),
  zipcode: z.string().nullable(),
  country_code: z.string().nullable(),
  country_name: z.string().nullable(),
  state_code: z.string().nullable(),
  state_text: z.string().nullable(),
  pickup_ready_in_minutes: z.number().nullable(),
  pickup_instructions: z.string().nullable(),
});

export type EmailStockLocation = z.infer<typeof EmailStockLocationSchema>;